backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.0
pydantic[email-validator]
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
redis==5.1.0
pika==1.3.2
pytest==8.3.0
httpx==0.27.0
pandas
slowapi
